+++ app/src/main/java/com/ahujafabrics/yarnit/Activity/CatalogItemView.java
@@ -1,9 +1,14 @@
 package com.ahujafabrics.yarnit.Activity;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
+import android.support.v4.content.ContextCompat;
+import android.support.v4.graphics.drawable.IconCompat;
+import android.support.v7.widget.CardView;
 import android.support.v7.widget.RecyclerView;
 import android.text.Editable;
 import android.text.TextWatcher;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -18,6 +23,8 @@ import java.util.List;
 
 public class CatalogItemView extends RecyclerView.Adapter<CatalogItemView.ViewHolder> {
 
+    private static final String TAG = "CatalogItemView";
+
     private Context context;
     private final List<ShadeCard> shadeGridValues;
     private OnCatalogItemClick ctlgItemClick;
@@ -31,16 +38,18 @@ public class CatalogItemView extends RecyclerView.Adapter<CatalogItemView.ViewHo
     public static class ViewHolder extends RecyclerView.ViewHolder{
         public TextView shadeLabel;
         public EditText qty;
+        public CardView shadeCard;
 
-        public MyCustomEditTextListener myCustomEditTextListener;
+        public QtyChangeEditTextListener qtyChangeEditTextListener;
 
-        public ViewHolder(View v, MyCustomEditTextListener myCustomEditTextListener){
+        public ViewHolder(View v, QtyChangeEditTextListener qtyChangeEditTextListener){
             super(v);
             shadeLabel = v.findViewById(R.id.shade);
             qty = v.findViewById(R.id.qty);
+            shadeCard = v.findViewById(R.id.shadCard);
 
-            this.myCustomEditTextListener = myCustomEditTextListener;
-            this.qty.addTextChangedListener(myCustomEditTextListener);
+            this.qtyChangeEditTextListener = qtyChangeEditTextListener;
+            this.qty.addTextChangedListener(qtyChangeEditTextListener);
         }
     }
 
@@ -51,19 +60,21 @@ public class CatalogItemView extends RecyclerView.Adapter<CatalogItemView.ViewHo
                 .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
 
         View v = inflater.inflate( R.layout.catalogitem , null);
-        ViewHolder vh = new ViewHolder(v,new MyCustomEditTextListener());
+        ViewHolder vh = new ViewHolder(v,new QtyChangeEditTextListener());
 
         return vh;
     }
 
-
     @Override
     public void onBindViewHolder(ViewHolder holder, int position){
 
-        holder.myCustomEditTextListener.updatePosition(holder.getAdapterPosition());
+        holder.qtyChangeEditTextListener.updatePosition(holder.getAdapterPosition());
 
         holder.shadeLabel.setText(shadeGridValues.get(position).getShade());
         holder.qty.setText(shadeGridValues.get(position).getQty());
+
+        if(!holder.qty.getText().toString().equals(""))
+            holder.shadeCard.setCardBackgroundColor(ContextCompat.getColor(context, R.color.cardBackground));
     }
 
     @Override
@@ -76,7 +87,8 @@ public class CatalogItemView extends RecyclerView.Adapter<CatalogItemView.ViewHo
         return shadeGridValues.size();
     }
 
-    private class MyCustomEditTextListener implements TextWatcher {
+    private class QtyChangeEditTextListener implements TextWatcher
+    {
         private int position;
 
         public void updatePosition(int position) {
@@ -92,11 +104,13 @@ public class CatalogItemView extends RecyclerView.Adapter<CatalogItemView.ViewHo
         public void onTextChanged(CharSequence charSequence, int i, int i2, int i3) {
             shadeGridValues.get(position).setQty(charSequence.toString());
             ctlgItemClick.onCatalogClick(shadeGridValues);
+
         }
 
         @Override
         public void afterTextChanged(Editable editable) {
             // no op
+
         }
     }
 }
