+++ app/src/main/java/com/ahujafabrics/yarnit/ViewOrder.java
@@ -72,11 +72,6 @@ public class ViewOrder extends AppCompatActivity {
 
         for(DataSnapshot ds : dataSnapshot.getChildren()){
             Order order = new Order();
-                    /*ds.child("orderID").getValue(String.class),
-                    ds.child("userID").getValue(String.class),
-                    ds.child("creationDate").getValue(long.class),
-                    ds.child("orderLineItems").getValue(List.class),
-                    ds.child("orderStatus").getValue(Order.OrderStatus.class)); */
 
             order.setOrderID(ds.child("orderID").getValue(String.class));
             order.setCreationDate(ds.child("creationDate").getValue(String.class));
@@ -86,15 +81,9 @@ public class ViewOrder extends AppCompatActivity {
             List<OrderItem> orderItemList = new ArrayList<>();
 
             for(DataSnapshot os : ds.child("orderLineItems").getChildren()){
-                /*OrderItem oi = new OrderItem(
-                        ds.child("orderLineItemID").getValue(Integer.class),
-                        ds.child("productType").getValue(String.class),
-                        ds.child("shadeId").getValue(String.class),
-                        ds.child("quantity").getValue(Integer.class)
-                ); */
+
                 OrderItem oi = new OrderItem();
                 oi.setOrderLineItemID(os.child("orderLineItemID").getValue(Integer.class));
-                oi.setProductType(os.child("productType").getValue(String.class));
                 oi.setQuantity(os.child("quantity").getValue(Integer.class));
                 oi.setShadeId(os.child("shadeId").getValue(String.class));
 
@@ -103,7 +92,6 @@ public class ViewOrder extends AppCompatActivity {
             GenericTypeIndicator<List<OrderItem>> genericOrderItemList = new GenericTypeIndicator<List<OrderItem>>() {};
 
             order.setOrderLineItems(ds.child("orderLineItems").getValue(genericOrderItemList));
-                    //ds.getValue(Order.class);
             ordersList.add(order);
         }
         // Initialize a new instance of RecyclerView Adapter instance
@@ -111,13 +99,6 @@ public class ViewOrder extends AppCompatActivity {
 
         // Set the adapter for RecyclerView
         mRecyclerView.setAdapter(viewOrderItemAdapter);
-    /*    OrderItem oi = new OrderItem(1, "Shantoon", "S/1", 4);
-        List<OrderItem> oiList = new ArrayList<>();
-        oiList.add(oi);
-        Order temp = new Order("1", "Ishwant", 2142019,
-                oiList, Order.OrderStatus.Submitted);
-        ordersList.add(temp); */
-
     }
 
 }
