+++ app/src/main/res/layout/catalogitem.xml
@@ -1,15 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:card_view="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
     android:layout_height="match_parent">
+
     <android.support.v7.widget.CardView
-        android:id="@+id/s"
+        android:id="@+id/shadCard"
         style="@style/shadeCardLayout">
 
         <LinearLayout style="@style/cardLinearLayout">
             <TextView
                 style="@style/ShadeCardTitle"
-                android:id="@+id/shade"/>
+                android:id="@+id/shade" />
             <View style="@style/shadeCardDividingLine" />
             <LinearLayout style="@style/qtyLayout">
                 <TextView style="@style/ShadeCardQtyText"
