+++ app/src/main/res/values/colors.xml
@@ -13,4 +13,5 @@
     <color name="pink">#fe104d</color>
     <color name="black_overlay">#66000000</color>
     <color name="white">#FFFFFF</color>
+    <color name="cardBackground">#EEE8AA</color>
 </resources>
